Add login modal to main-page
The modal appears when link "Συνδεση" is clicked.
the modal asks for the users email and provides
alternative options to login with google,facebook
or apple.Lastly the modal provides a link to the
terms and conditions of the website.
The link doesnt work yet plus the modal is not
centered.